split FileInput from Input, change some props
